Rename `dash_component_editor` to better represent new structure
